+++ ReactRedux/src/components/Main.js
@@ -7,9 +7,7 @@ var Main = React.createClass({
             <div>
                 <h1>이모콘 경품 추첨</h1>
                 <p>이모콘에서 경품 추첨을 위해 사용하기 위해 작성된 소프트웨어입니다.</p>
-                <p>
-                    다음 순서로 경품 추첨을 하게 됩니다.
-                </p>
+                <p>다음 순서로 경품 추첨을 하게 됩니다.</p>
                 <ol>
                     <li>참가자 목록이 담긴 CSV 파일을 읽어들입니다.</li>
                     <li>경품을 입력합니다.</li>
@@ -17,10 +15,7 @@ var Main = React.createClass({
                     <li>당첨자들을 확인합니다.</li>
                 </ol>
                 <p>
-                    사용해주셔서 감사합니다.
-                </p>
-                <p>
-                    경품 추첨을 시작해봅니다. <Link to="/uploadCSV">여기를</Link> 눌러주세요.
+                    경품 추첨을 시작하려면 <Link to="/uploadCSV">여기</Link>를 눌러주세요.
                 </p>
             </div>
         );
